Add files via upload
Hi! I am William Beamer, and this is my latest Linux operating system, BeeOS 1.0! To install, be sure to load all of these files onto a USB drive, and then insert the drive into your desired device! This OS is mostly Linux-based, but it is mostly inspired by PearOS and Ubuntu! Happy Coding!